Fix variable shadowing in new_path_with_id
In the current version, whenever a scid is None when
calling new_with_path_id, the reset token associated to a
dcid is also set to None, even if it had a value before.
This leads to errors in the verification of the CID
if a NEW_PATH_CONNECTION_ID is retransmitted (e.g.: because
it was lost), as reset tokens differs.

This patch fixes this issue by deleting the variable shadowing.